requirements.txt
fastapi
uvicorn
websockets
jinja2
psycopg2-binary
python-dotenv
langchain
langchain-openai
langgraph
langchain-community
huggingface-hub
sentence-transformers
chromadb
twilio
reportlab
pydantic